--- src/stories/introduction.stories.mdx
@@ -1,60 +1,0 @@
-import { Meta, Story, Canvas, Description, ArgsTable } from '@storybook/addon-docs/blocks';
-import Lookingglass from '../lib/components/Lookingglass';
-import Flex from '../lib/components/Flex';
-
-<Meta title="Introduction/Introduction" />
-
-# Lookingglass
-
-<hr />
-
-<Lookingglass div fontSize="1" marginBottom="5">
-    Lookingglass is a UI component library built with React and styled in Sass.
-</Lookingglass>
-
-## Getting Started
-
-Install the npm package below into your project,
-which contains everything you need to get started.
-
-```
-    npm install lookingglass-react
-```
-
-#### Using Components
-
-You can use a component by using an import in the below format.
-If you are unsure of the import you need, then visit the component's main page
-
-```js
-import Alert from 'lookingglass-react/Alert';
-```
-
-#### Importing Styles
-
-In order to setup the stylesheet, you will need Sass configured in your project.
-To apply all styles, from your Sass file you can `@use` the below entry point.
-
-```scss
-@use 'lookingglass-react/sass/index';
-```
-
-Or if you are only using a few components, only import the styles you are using
-
-```scss
-@use 'lookingglass-react/sass/components/alerts';
-```
-
-#### Global Styles
-
-The following reset should be defined in order for component sizing to work as expected.
-This has already been included in the main SASS entry point.
-You will need to include this in your stylesheet if you are not using all styles.
-
-```css
-*,
-*::before,
-*::after {
-    box-sizing: border-box;
-}
-```
