--- src/stories/typography.stories.js
@@ -1,42 +1,0 @@
-import Lookingglass from '../lib/components/Lookingglass';
-
-export const Headers = (args) => {
-    return (
-        <>
-            <h1>This is a H1 Heading</h1>
-            <h2>This is a H2 Heading</h2>
-            <h3>This is a H3 Heading</h3>
-            <h4>This is a H4 Heading</h4>
-            <h5>This is a H5 Heading</h5>
-            <h6>This is a H6 Heading</h6>
-        </>
-    );
-};
-
-export const Paragraph = (args) => (
-    <div>
-        <p>
-            Lorem ipsum, dolor sit amet consectetur adipisicing elit. Neque, officiis! Neque, nostrum natus ipsum dicta
-            perferendis porro nisi fugit saepe ratione voluptas aperiam facere ea, dolorum consequatur? Ipsum optio aut
-            doloremque fugiat commodi odit nobis delectus magnam! Fuga ratione voluptates doloremque vero nulla,
-            voluptate non cupiditate iste, quod, qui placeat nobis minima reprehenderit sed? Nulla, asperiores incidunt
-            amet laboriosam molestiae ducimus neque molestias odit laudantium delectus inventore maiores explicabo quod
-            cum dolore, ratione iste numquam ut ipsam. Corporis harum adipisci iure quidem laudantium voluptatem
-            distinctio rerum veritatis quam deserunt facere natus hic, temporibus explicabo mollitia at similique
-            consequatur! Cupiditate, molestiae!
-        </p>
-        <p>
-            Lorem ipsum, dolor sit amet consectetur adipisicing elit. Neque, officiis! Neque, nostrum natus ipsum dicta
-            perferendis porro nisi fugit saepe ratione voluptas aperiam facere ea, dolorum consequatur? Ipsum optio aut
-            doloremque fugiat commodi odit nobis delectus magnam! Fuga ratione voluptates doloremque vero nulla,
-            voluptate non cupiditate iste, quod, qui placeat nobis minima reprehenderit sed? Nulla, asperiores incidunt
-            amet laboriosam molestiae ducimus neque molestias odit laudantium delectus inventore maiores explicabo quod
-            cum dolore, ratione iste numquam ut ipsam. Corporis harum adipisci iure quidem laudantium voluptatem
-            distinctio rerum veritatis quam deserunt facere natus hic, temporibus explicabo mollitia at similique
-            consequatur! Cupiditate, molestiae!
-        </p>
-        <p>
-            Some Example Text with a <a href="#">Test Link</a>
-        </p>
-    </div>
-);
